+++ Common/SargeStoreDomain/Models/Cart.cs
@@ -7,6 +7,6 @@ namespace SargeStoreDomain.Models
     {
         public List<CartItem> Items { get; set; } = new List<CartItem>();
 
-        public int ItemCount => Items?.Sum(item => item.Quantity) ?? 0;
+        public int ItemsCount => Items?.Sum(item => item.Quantity) ?? 0;
     }
 }
