+++ Services/SargeStore.Interfaces/Services/ICartStore.cs
@@ -1,0 +1,9 @@
+﻿using SargeStoreDomain.Models;
+
+namespace SargeStore.Interfaces.Services
+{
+    public interface ICartStore
+    {
+        Cart Cart { get; set; }
+    }
+}
